+++ src/main/java/com/jsorrell/carpetskyadditions/advancements/predicates/SkyAdditionsLootItemConditions.java
@@ -8,8 +8,10 @@ import net.minecraft.world.level.storage.loot.predicates.LootItemCondition;
 import net.minecraft.world.level.storage.loot.predicates.LootItemConditionType;
 
 public class SkyAdditionsLootItemConditions {
-    public static final LootItemConditionType DESERT_PYRAMID_EXISTS =
-            register("desert_pyramid_exists", new DesertPyramidExistsCondition.Serializer());
+    public static final LootItemConditionType LOCATION_CHECK =
+            register("location_check", new SkyAdditionsLocationCheck.Serializer());
+    public static final LootItemConditionType ENTITY_PROPERTIES =
+            register("entity_properties", new SkyAdditionsLootItemEntityPropertyCondition.Serializer());
 
     private static LootItemConditionType register(
             String registryName, Serializer<? extends LootItemCondition> serializer) {