+++ src/main/java/com/jsorrell/carpetskyadditions/mixin/BlocksMixin.java
@@ -1,0 +1,21 @@
+package com.jsorrell.carpetskyadditions.mixin;
+
+import com.jsorrell.carpetskyadditions.helpers.CoralSpreader;
+import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.state.BlockBehaviour;
+import org.spongepowered.asm.mixin.Mixin;
+import org.spongepowered.asm.mixin.injection.At;
+import org.spongepowered.asm.mixin.injection.Redirect;
+import org.spongepowered.asm.mixin.injection.Slice;
+
+@Mixin(Blocks.class)
+public abstract class BlocksMixin {
+    @Redirect(
+            method = "<clinit>",
+            slice = @Slice(from = @At(value = "CONSTANT", args = "stringValue=calcite")),
+            at = @At(value = "NEW", args = "class=net/minecraft/world/level/block/Block", ordinal = 0))
+    private static Block registerCustomCalcite(BlockBehaviour.Properties properties) {
+        return new CoralSpreader.CustomCalciteBlock(properties);
+    }
+}