+++ src/main/java/com/jsorrell/carpetskyadditions/settings/SkyAdditionsSettings.java
@@ -250,6 +250,10 @@ public class SkyAdditionsSettings {
     @SkyAdditionsSetting("true")
     public static boolean spreadingSmallDripleaves = false;
 
+    @Rule(categories = FEATURE)
+    @SkyAdditionsSetting("true")
+    public static boolean spreadingCoral = false;
+
     /* Wandering Trader Spawn Chance */
     public static class WanderingTraderSpawnChanceValidator extends Validator<Double> {
         @Override